+++ source/isaaclab_newton/test/sensors/__init__.py
@@ -2,10 +2,3 @@
 # All rights reserved.
 #
 # SPDX-License-Identifier: BSD-3-Clause
-
-__all__ = [
-    "newton_physics_replicate",
-    "newton_visualizer_prebuild",
-]
-
-from .newton_replicate import newton_physics_replicate, newton_visualizer_prebuild