+++ source/isaaclab_tasks/isaaclab_tasks/manager_based/manipulation/dexsuite/config/kuka_allegro/agents/rsl_rl_ppo_cfg.py
@@ -8,34 +8,37 @@ from dataclasses import MISSING
 from isaaclab.utils import configclass
 
 from isaaclab_rl.rsl_rl import (
-    RslRlActorCriticCNNCfg,
+    RslRlCNNModelCfg,
+    RslRlMLPModelCfg,
     RslRlOnPolicyRunnerCfg,
-    RslRlPpoActorCriticCfg,
     RslRlPpoAlgorithmCfg,
 )
 
 from isaaclab_tasks.utils import PresetCfg
 
-STATE_POLICY_CFG = RslRlPpoActorCriticCfg(
-    init_noise_std=1.0,
-    actor_obs_normalization=True,
-    critic_obs_normalization=True,
-    actor_hidden_dims=[512, 256, 128],
-    critic_hidden_dims=[512, 256, 128],
+STATE_POLICY_CFG = RslRlMLPModelCfg(
+    distribution_cfg=RslRlMLPModelCfg.GaussianDistributionCfg(init_std=1.0),
+    obs_normalization=True,
+    hidden_dims=[512, 256, 128],
     activation="elu",
 )
 
-CNN_POLICY_CFG = RslRlActorCriticCNNCfg(
-    init_noise_std=1.0,
-    actor_obs_normalization=True,
-    critic_obs_normalization=True,
-    actor_hidden_dims=[512, 256, 128],
-    critic_hidden_dims=[512, 256, 128],
-    actor_cnn_cfg=RslRlActorCriticCNNCfg.CNNCfg(
-        output_channels=[16, 32, 64],
-        kernel_size=[3, 3, 3],
+
+STATE_CRITIC_CFG = RslRlMLPModelCfg(
+    obs_normalization=True,
+    hidden_dims=[512, 256, 128],
+    activation="elu",
+)
+
+CNN_POLICY_CFG = RslRlCNNModelCfg(
+    obs_normalization=True,
+    hidden_dims=[512, 256, 128],
+    distribution_cfg=RslRlCNNModelCfg.GaussianDistributionCfg(init_std=1.0),
+    cnn_cfg=RslRlCNNModelCfg.CNNCfg(
+        output_channels=[16, 32],
+        kernel_size=[3, 3],
         activation="elu",
-        max_pool=[True, True, True],
+        max_pool=[True, True],
         norm="batch",
         global_pool="avg",
     ),
@@ -64,10 +67,10 @@ class DexsuiteKukaAllegroPPOBaseRunnerCfg(RslRlOnPolicyRunnerCfg):
     num_steps_per_env = 32
     max_iterations = 15000
     save_interval = 250
-    empirical_normalization = True
     experiment_name = (MISSING,)  # type: ignore
     obs_groups = (MISSING,)  # type: ignore
-    policy = (MISSING,)  # type: ignore
+    actor = (MISSING,)  # type: ignore
+    critic = (MISSING,)  # type: ignore
     algorithm = MISSING  # type: ignore
 
 
@@ -75,24 +78,27 @@ class DexsuiteKukaAllegroPPOBaseRunnerCfg(RslRlOnPolicyRunnerCfg):
 class DexsuiteKukaAllegroPPORunnerCfg(PresetCfg):
     default = DexsuiteKukaAllegroPPOBaseRunnerCfg().replace(
         experiment_name="dexsuite_kuka_allegro",
-        obs_groups={"policy": ["policy", "proprio", "perception"], "critic": ["policy", "proprio", "perception"]},
-        policy=STATE_POLICY_CFG,
+        obs_groups={"actor": ["policy", "proprio", "perception"], "critic": ["policy", "proprio", "perception"]},
+        actor=STATE_POLICY_CFG,
+        critic=STATE_CRITIC_CFG,
         algorithm=ALGO_CFG,
     )
 
     single_camera = DexsuiteKukaAllegroPPOBaseRunnerCfg().replace(
         experiment_name="dexsuite_kuka_allegro_single_camera",
-        obs_groups={"policy": ["policy", "proprio", "base_image"], "critic": ["policy", "proprio", "perception"]},
-        policy=CNN_POLICY_CFG,
+        obs_groups={"actor": ["policy", "proprio", "base_image"], "critic": ["policy", "proprio", "perception"]},
+        actor=CNN_POLICY_CFG,
+        critic=STATE_CRITIC_CFG,
         algorithm=ALGO_CFG.replace(num_mini_batches=16),
     )
 
     duo_camera = DexsuiteKukaAllegroPPOBaseRunnerCfg().replace(
         experiment_name="dexsuite_kuka_allegro_duo_camera",
         obs_groups={
-            "policy": ["policy", "proprio", "base_image", "wrist_image"],
+            "actor": ["policy", "proprio", "base_image", "wrist_image"],
             "critic": ["policy", "proprio", "perception"],
         },
-        policy=CNN_POLICY_CFG,
+        actor=CNN_POLICY_CFG,
+        critic=STATE_CRITIC_CFG,
         algorithm=ALGO_CFG.replace(num_mini_batches=16),
     )